OD-10215 Remove reindexCourses job - never used. Do not reindex classes for tests - reindex courses only.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@53219 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ common/solr-index/src/main/java/ish/oncourse/solr/reindex/ReindexCourses.java
@@ -19,10 +19,11 @@ import java.util.concurrent.Callable;
 import java.util.function.Function;
 
 public class ReindexCourses extends ReindexCollection<SCourse> {
-
+	private boolean reindexClasses = true;
 
 	public ReindexCourses(ObjectContext objectContext, SolrClient solrClient) {
 		this(objectContext, solrClient, s -> new String[0]);
+		reindexClasses = false;
 	}
 	
 	
@@ -44,16 +45,20 @@ public class ReindexCourses extends ReindexCollection<SCourse> {
 
 	public UpdateResponse addBean(SCourse sCourse) throws IOException, SolrServerException {
 		total.incrementAndGet();
-		solrClient.deleteByQuery(SolrCollection.classes.name(), String.format("courseId:%s", sCourse.getId()));
-		
-		for (SCourseClass c : sCourse.getClasses()) {
-			solrClient.addBean(SolrCollection.classes.name(), c);
+		if (reindexClasses) {
+			solrClient.deleteByQuery(SolrCollection.classes.name(), String.format("courseId:%s", sCourse.getId()));
+
+			for (SCourseClass c : sCourse.getClasses()) {
+				solrClient.addBean(SolrCollection.classes.name(), c);
+			}
 		}
 		return solrClient.addBean(SolrCollection.courses.name(), sCourse);
 	}
 
 	protected void commit() throws IOException, SolrServerException {
 		super.commit();
-		solrClient.commit(SolrCollection.classes.name());
+		if (reindexClasses) {
+			solrClient.commit(SolrCollection.classes.name());
+		}
 	}
 }
